controlnet_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
torch>=2.1.0
diffusers==0.27.2
transformers>=4.39.0
accelerate>=0.28.0
xformers>=0.0.23
opencv-python>=4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0
httpx>=0.27.0
python-dotenv>=1.0.1
